+++ tests/unit/console_results_test.sh
@@ -602,7 +602,7 @@ function test_print_successful_test_output_no_args() {
 
   assert_matches \
     "✓ Passed.*$test_name.*12ms" \
-    "$(bashunit::console_results::print_successful_test "$test_name" "12")"
+    "$(BASHUNIT_SHOW_EXECUTION_TIME=true bashunit::console_results::print_successful_test "$test_name" "12")"
 
   export BASHUNIT_SIMPLE_OUTPUT=$original_simple_output
 }
@@ -617,7 +617,7 @@ function test_print_successful_test_output_with_args() {
 
   assert_matches \
     "✓ Passed.*$test_name \('$data'\).*12ms" \
-    "$(bashunit::console_results::print_successful_test "$test_name" "12" "$data")"
+    "$(BASHUNIT_SHOW_EXECUTION_TIME=true bashunit::console_results::print_successful_test "$test_name" "12" "$data")"
 
   export BASHUNIT_SIMPLE_OUTPUT=$original_simple_output
 }
@@ -631,7 +631,7 @@ function test_print_successful_test_output_in_seconds() {
 
   assert_matches \
     "✓ Passed.*$test_name.*5.12s" \
-    "$(bashunit::console_results::print_successful_test "$test_name" "5123")"
+    "$(BASHUNIT_SHOW_EXECUTION_TIME=true bashunit::console_results::print_successful_test "$test_name" "5123")"
 
   export BASHUNIT_SIMPLE_OUTPUT=$original_simple_output
 }
@@ -645,7 +645,7 @@ function test_print_successful_test_output_in_minutes() {
 
   assert_matches \
     "✓ Passed.*$test_name.*1m 3s" \
-    "$(bashunit::console_results::print_successful_test "$test_name" "63000")"
+    "$(BASHUNIT_SHOW_EXECUTION_TIME=true bashunit::console_results::print_successful_test "$test_name" "63000")"
 
   export BASHUNIT_SIMPLE_OUTPUT=$original_simple_output
 }
@@ -659,7 +659,7 @@ function test_print_successful_test_output_in_minutes_exact() {
 
   assert_matches \
     "✓ Passed.*$test_name.*2m 0s" \
-    "$(bashunit::console_results::print_successful_test "$test_name" "120000")"
+    "$(BASHUNIT_SHOW_EXECUTION_TIME=true bashunit::console_results::print_successful_test "$test_name" "120000")"
 
   export BASHUNIT_SIMPLE_OUTPUT=$original_simple_output
 }